rename: Move chat classes into dedicated api, domain, and repository packages
