--- src/main/java/jsbh/Jusangbokhap/chat/chatmessage/ChatMessageService.java
@@ -1,18 +1,0 @@
-package jsbh.Jusangbokhap.chat.chatmessage;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class ChatMessageService {
-
-	private final ChatMessageRepository chatMessageRepository;
-
-	public List<ChatMessage> getMessages(Long roomId) {
-		return chatMessageRepository.findByRoomIdOrderByTimestampAsc(roomId);
-	}
-}
