+++ src/main/java/jsbh/Jusangbokhap/common/config/WebSocketConfig.java
@@ -5,7 +5,7 @@ import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
 
-import jsbh.Jusangbokhap.chat.ChatWebSocketHandler;
+import jsbh.Jusangbokhap.api.chat.ChatWebSocketHandler;
 import lombok.RequiredArgsConstructor;
 
 @Configuration
