+++ src/main/java/jsbh/Jusangbokhap/domain/chat/repository/jpa/ChatRoomRepository.java
@@ -1,4 +1,4 @@
-package jsbh.Jusangbokhap.chat.chatroom;
+package jsbh.Jusangbokhap.domain.chat.repository.jpa;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,6 +6,8 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+
+import jsbh.Jusangbokhap.domain.chat.entity.jpa.ChatRoom;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
 	@Query("SELECT c FROM ChatRoom c WHERE (c.participant1Id = :participant1Id AND c.participant2Id = :participant2Id) OR (c.participant1Id = :participant2Id AND c.participant2Id = :participant1Id)")
