+++ src/test/java/jsbh/Jusangbokhap/chat/chatroom/ChatRoomRepositoryTest.java
@@ -6,7 +6,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.ActiveProfiles;
+
+import jsbh.Jusangbokhap.domain.chat.entity.jpa.ChatRoom;
+import jsbh.Jusangbokhap.domain.chat.repository.jpa.ChatRoomRepository;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
